+++ examples/cloud/from_documents.ts
@@ -16,28 +16,25 @@ async function main() {
 
   const index = await LlamaCloudIndex.fromDocuments({
     documents: [document],
-    name: "test",
-    projectName: "default",
+    name: "test-pipeline",
+    projectName: "Default",
     apiKey: process.env.LLAMA_CLOUD_API_KEY,
     baseUrl: process.env.LLAMA_CLOUD_BASE_URL,
   });
 
   const queryEngine = index.asQueryEngine({
-    denseSimilarityTopK: 5,
+    similarityTopK: 5,
   });
 
   const rl = readline.createInterface({ input, output });
 
   while (true) {
     const query = await rl.question("Query: ");
-    const stream = await queryEngine.query({
+    const response = await queryEngine.query({
       query,
-      stream: true,
     });
-    console.log();
-    for await (const chunk of stream) {
-      process.stdout.write(chunk.response);
-    }
+
+    console.log(response.toString());
   }
 }