+++ packages/cloud/openapi-ts.config.ts
@@ -1,0 +1,15 @@
+import { defineConfig } from "@hey-api/openapi-ts";
+
+export default defineConfig({
+  // you can download this file to get the latest version of the OpenAPI document
+  // @link https://api.cloud.llamaindex.ai/api/openapi.json
+  input: "./openapi.json",
+  output: {
+    path: "./src/client",
+    format: "prettier",
+    lint: "eslint",
+  },
+  types: {
+    enums: "javascript",
+  },
+});