+++ packages/llamaindex/src/cloud/utils.ts
@@ -1,7 +1,7 @@
-import type { PlatformApiClient } from "@llamaindex/cloud";
+import { OpenAPI, Service } from "@llamaindex/cloud/api";
 import { getEnv } from "@llamaindex/env";
-import type { ClientParams } from "./types.js";
-import { DEFAULT_BASE_URL } from "./types.js";
+import type { ClientParams } from "./constants.js";
+import { DEFAULT_BASE_URL } from "./constants.js";
 
 function getBaseUrl(baseUrl?: string): string {
   return baseUrl ?? getEnv("LLAMA_CLOUD_BASE_URL") ?? DEFAULT_BASE_URL;
@@ -11,18 +11,17 @@ export function getAppBaseUrl(baseUrl?: string): string {
   return getBaseUrl(baseUrl).replace(/api\./, "");
 }
 
-export async function getClient({
+export function initService({
   apiKey,
   baseUrl,
-}: ClientParams = {}): Promise<PlatformApiClient> {
-  // Get the environment variables or use defaults
-  baseUrl = getBaseUrl(baseUrl);
-  apiKey = apiKey ?? getEnv("LLAMA_CLOUD_API_KEY");
+}: ClientParams = {}): typeof Service {
+  OpenAPI.TOKEN = apiKey ?? getEnv("LLAMA_CLOUD_API_KEY");
+  OpenAPI.BASE = getBaseUrl(baseUrl);
+  if (!OpenAPI.TOKEN) {
+    throw new Error(
+      "API Key is required for LlamaCloudIndex. Please pass the apiKey parameter",
+    );
+  }
 
-  const { PlatformApiClient } = await import("@llamaindex/cloud");
-
-  return new PlatformApiClient({
-    token: apiKey,
-    environment: baseUrl,
-  });
+  return Service;
 }