+++ packages/llamaindex/src/llm/gemini/utils.ts
@@ -1,17 +1,23 @@
-import { type Content as GeminiMessageContent } from "@google/generative-ai";
+import {
+  type FunctionCall,
+  type Content as GeminiMessageContent,
+} from "@google/generative-ai";
 
 import { type GenerateContentResponse } from "@google-cloud/vertexai";
+import type { BaseTool } from "../../types.js";
 import type {
   ChatMessage,
-  MessageContent,
   MessageContentImageDetail,
   MessageContentTextDetail,
   MessageType,
+  ToolCallLLMMessageOptions,
 } from "../types.js";
 import { extractDataUrlComponents } from "../utils.js";
 import type {
   ChatContext,
   FileDataPart,
+  FunctionDeclaration,
+  FunctionDeclarationSchema,
   GeminiChatParamsNonStreaming,
   GeminiChatParamsStreaming,
   GeminiMessageRole,
@@ -104,7 +110,8 @@ export const cleanParts = (
         part.text?.trim() ||
         part.inlineData ||
         part.fileData ||
-        part.functionCall,
+        part.functionCall ||
+        part.functionResponse,
     ),
   };
 };
@@ -115,8 +122,21 @@ export const getChatContext = (
   // Gemini doesn't allow:
   // 1. Consecutive messages from the same role
   // 2. Parts that have empty text
+  const fnMap = params.messages.reduce(
+    (result, message) => {
+      if (message.options && "toolCall" in message.options)
+        message.options.toolCall.forEach((call) => {
+          result[call.id] = call.name;
+        });
+
+      return result;
+    },
+    {} as Record<string, string>,
+  );
   const messages = GeminiHelper.mergeNeighboringSameRoleMessages(
-    params.messages.map(GeminiHelper.chatMessageToGemini),
+    params.messages.map((message) =>
+      GeminiHelper.chatMessageToGemini(message, fnMap),
+    ),
   ).map(cleanParts);
 
   const history = messages.slice(0, -1);
@@ -124,6 +144,23 @@ export const getChatContext = (
   return {
     history,
     message,
+  };
+};
+
+export const mapBaseToolToGeminiFunctionDeclaration = (
+  tool: BaseTool,
+): FunctionDeclaration => {
+  const parameters: FunctionDeclarationSchema = {
+    type: tool.metadata.parameters?.type.toUpperCase(),
+    properties: tool.metadata.parameters?.properties,
+    description: tool.metadata.parameters?.description,
+    required: tool.metadata.parameters?.required,
+  };
+
+  return {
+    name: tool.metadata.name,
+    description: tool.metadata.description,
+    parameters,
   };
 };
 
@@ -177,7 +214,40 @@ export class GeminiHelper {
       );
   }
 
-  public static messageContentToGeminiParts(content: MessageContent): Part[] {
+  public static messageContentToGeminiParts({
+    content,
+    options = undefined,
+    fnMap = undefined,
+  }: Pick<ChatMessage<ToolCallLLMMessageOptions>, "content" | "options"> & {
+    fnMap?: Record<string, string>;
+  }): Part[] {
+    if (options && "toolResult" in options) {
+      if (!fnMap) throw Error("fnMap must be set");
+      const name = fnMap[options.toolResult.id];
+      if (!name)
+        throw Error(
+          `Could not find the name for fn call with id ${options.toolResult.id}`,
+        );
+
+      return [
+        {
+          functionResponse: {
+            name,
+            response: {
+              result: options.toolResult.result,
+            },
+          },
+        },
+      ];
+    }
+    if (options && "toolCall" in options) {
+      return options.toolCall.map((call) => ({
+        functionCall: {
+          name: call.name,
+          args: call.input,
+        } as FunctionCall,
+      }));
+    }
     if (typeof content === "string") {
       return [{ text: content }];
     }
@@ -197,11 +267,35 @@ export class GeminiHelper {
   }
 
   public static chatMessageToGemini(
-    message: ChatMessage,
+    message: ChatMessage<ToolCallLLMMessageOptions>,
+    fnMap: Record<string, string>, // mapping of fn call id to fn call name
   ): GeminiMessageContent {
     return {
       role: GeminiHelper.ROLES_TO_GEMINI[message.role],
-      parts: GeminiHelper.messageContentToGeminiParts(message.content),
+      parts: GeminiHelper.messageContentToGeminiParts({ ...message, fnMap }),
     };
+  }
+}
+
+/**
+ * Returns functionCall of first candidate.
+ * Taken from https://github.com/google-gemini/generative-ai-js/ to be used with
+ * vertexai as that library doesn't include it
+ */
+export function getFunctionCalls(
+  response: GenerateContentResponse,
+): FunctionCall[] | undefined {
+  const functionCalls: FunctionCall[] = [];
+  if (response.candidates?.[0].content?.parts) {
+    for (const part of response.candidates?.[0].content?.parts) {
+      if (part.functionCall) {
+        functionCalls.push(part.functionCall);
+      }
+    }
+  }
+  if (functionCalls.length > 0) {
+    return functionCalls;
+  } else {
+    return undefined;
   }
 }